Refactor[bmq]: move bmqu::TempFile -> bmqtst::TempFile
