+++ src/groups/bmq/bmqtst/doc/bmqtst.txt
@@ -9,12 +9,13 @@ writing test drivers.
 
 /Hierarchical Synopsis
 /---------------------
- The bmqtst' package currently has 4 components having 1 level of physical
+ The bmqtst' package currently has 5 components having 1 level of physical
  dependency.  The list below shows the hierarchical ordering of the components.
 ..
   1. bmqtst_blobtestutil
      bmqtst_mocktime
      bmqtst_scopedlogobserver
+     bmqtst_tempfile
      bmqtst_testhelper
 ..
 
@@ -26,5 +27,7 @@ writing test drivers.
 :      Provide a mock of the time accessors usable with 'bmqu::Time'.
 : 'bmqtst_scopedlogobserver':
 :      Provide a scoped tester implementation of the log observer protocol.
+: 'bmqtst_tempfile':
+:      Provide a guard for creating a temporary file.
 : 'bmqtst_testhelper':
 :      Provide macros and utilities to assist in writing test drivers.
