+++ src/groups/bmq/bmqtst/package/bmqtst.mem
@@ -2,4 +2,5 @@ bmqtst_blobtestutil
 bmqtst_mocktime
 bmqtst_scopedlogobserver
 bmqtst_table
+bmqtst_tempfile
 bmqtst_testhelper
